power: smb5-lib: Populate the soft jeita thresholds
Currently if the soft jeita and soft jeita hysteresis thresholds
are not configured in the device tree file, the corresponding
jeita registers are programmed to '0', which is not expected
behaviour. Fix it by populating the soft jeita thresholds with
the corresponding register values in the absence of soft jeita
and soft jeita hysteresis properties in device tree file.

Change-Id: I985121ec1b8897549b3256a5de1cdf2effee5921
Signed-off-by: Kiran Gunda <kgunda@codeaurora.org>